+++ packages/web_ui/src/components/SiteLayout.jsx
@@ -17,11 +17,14 @@ import UsersPage from "./UsersPage";
 import UserViewPage from "./UserViewPage";
 import RolesPage from "./RolesPage";
 import RoleViewPage from "./RoleViewPage";
+import PluginsPage from "./PluginsPage";
+import PluginViewPage from "./PluginViewPage";
 
 const pages = [
 	{ path: "/instances/:id/view", sidebarPath: "/instances", content: <InstanceViewPage /> },
 	{ path: "/users/:name/view", sidebarPath: "/users", content: <UserViewPage /> },
 	{ path: "/roles/:id/view", sidebarPath: "/roles", content: <RoleViewPage /> },
+	{ path: "/plugins/:name/view", sidebarPath: "/plugins", content: <PluginViewPage /> },
 ];
 
 const sidebar = [
@@ -30,6 +33,7 @@ const sidebar = [
 	{ name: "Instances", path: "/instances", content: <InstancesPage />},
 	{ name: "Users", path: "/users", content: <UsersPage /> },
 	{ name: "Roles", path: "/roles", content: <RolesPage /> },
+	{ name: "Plugins", path: "/plugins", content: <PluginsPage /> },
 ];
 
 
@@ -75,7 +79,9 @@ export default function SiteLayout(props) {
 				<Switch>
 					{sidebar.map(({path, content}) => <Route exact path={path} key={path}>
 						<SetSidebar path={path} />
-						{content}
+						<ErrorBoundary Component={ErrorPage}>
+							{content}
+						</ErrorBoundary>
 					</Route>)}
 					{pages.map(page => <Route exact path={page.path} key={page.path}>
 						<SetSidebar path={page.sidebarPath} />